requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.2
numpy==1.26.1
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2